2.77
Added line to check for custom file name of script, so that way on update the script will retain the name given by the user. Updated verbiage in the beginning of the script and within the Update Script itself. Modified some file names, so they were deleted as to be re-uploaded.