2.90
Added in testing repository packages for Raspbian. Will not install if not running Raspbian Stretch. This is required because there are many changes to the Drawpile source code by QT that cannot be easily patched or changed. This should provide a fix for a while and prevent compiling problems. Test repository packages are generally speaking stable as they are candidates to be moved to the stable branch.