2.94
Minor update. Includes some changes from major script update work in progress to present better options for locations that actually exist on the server in use. Also provides a new script function for ensuring specific packages are installed, in this case, curl for freeDNS as it won't work without this package. However, these changes are inconsistent and won't show up in the server management script because of the length of time it would take to properly edit the scripts to do so. This will be fixed with the next major script release (hopefully soon).