2.95
Minor change to the NGINX configuration required as a result of Drawpile 2.14's release: "Note for people running servers: the admin API root path has changed! The paths are now prefixed with "/api", so "/status/" becomes "/api/status/". This change is to better support the new work-in-progress web admin frontend."

https://drawpile.net/news/release-2.1.14/