NGINX Fedora
NGINX not working, because Fedora does something non-standard and is content to leave things broken (because they can). Most NGINX installs do not create a default server for you, they let you do one or disable it via a separate config file. Fedora jams it into the main nginx.conf file, which is not the way everyone else does it.
https://bugzilla.redhat.com/show_bug.cgi?id=1220094

This removes the default server config, so the installation can be scripted more easily.